requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
pandas
faker
chromadb
python-dotenv
sqlparse
streamlit
requests
matplotlib
seaborn
